// TaskFlow.Web/src/api/client/types.gen.ts
export type CreateTaskItemDto = {
    title?: string;
    description?: null | string;
    status?: Status;
    isComplete?: boolean;
    priority?: Priority;
    dueDate?: null | string;
};

export type NoteResponseDto = {
    id?: number | string;
    content: string;
    taskItemId?: number | string;
    createdAt?: string;
    updatedAt?: null | string;
};

export type Priority = 'low' | 'medium' | 'high';

export type Status = 'draft' | 'todo' | 'completed';

export type TaskItemResponseDto = {
    id?: number | string;
    title: string;
    description?: null | string;
    isComplete?: boolean;
    dueDate?: null | string;
    status?: string;
    priority?: string;
};